50/50 joker logic, css, images for the jokers
