--- src/components/Jokers/50-50.module.css
@@ -1,16 +1,0 @@
-.jokers {
-    width: 5em ;
-    height: 5em;
-    display: flex;
-    flex-direction: column;
-   margin: 0 auto;
-    padding: 2em;
-}
-
-.fifty{
-    background-image: url("assets/images/Classic5050.jpg");
-    background-size: contain;
-    background-repeat: no-repeat;
-    align-items: center;
-   
-}
