+++ src/pages/Game/GameContent.js
@@ -4,7 +4,7 @@ import Question from "components/Question/Question";
 import Answers from "components/Answers/Answers";
 import Timer from "components/Timer/Timer";
 import Player from "components/Sound/Sound";
-//import RemoveTwoRandomElements from "components/Jokers/50-50";
+import Jokers from "components/Jokers/Jokers";
 
 import styles from "./GameContent.module.css";
 
@@ -41,9 +41,9 @@ const GameContent = ({ currentQuestion, answeredQuestionsCount }) => {
 
   return (
     <div className={styles.background}>
-      {/* <RemoveTwoRandomElements /> */}
       <Player />
       <Timer />
+      <Jokers />
       <div className={styles.gameContainer}>
         <Question
           question={currentQuestion.question}
